Changed all of the links for the new website.
So the easiest way to do this was to go into the WYSIWYG editor and add
one link for each page next to it then get the source code then. Modify
the text for the link to what it used to say, then modify the old links
for other items to the new links.